refactor(web,listings): apply tech-review suggestions for the #810 blocker
Self-review follow-ups (all SUGGESTION-level), no behaviour change to the
blocker logic:

- a11y: make the review readiness summary a polite live region and the
  missing-product-params nudge a `role="status"` region, so their async
  appearance/updates are announced.
- style: align the nudge font-size to the canonical type scale
  (`0.8125rem`) instead of an off-scale `12.5px`.
- simplify: drop the hand-rolled signature memo in
  `useBulkRequiredProductParams` — the wizard's reconcile effect already
  short-circuits on an unchanged blocker set, so a fresh Map per render is
  harmless.
- consolidate: move the pure blocker-derivation helpers
  (`categoryResultFor`, `selectBulkProductCardId`, `recomputeRowBlockers`)
  out of `bulk-wizard.tsx` into `bulk-policy.ts`, where `computeBlockers`
  lives. Tests import them from their new home.
