+++ apps/web/src/index.css
@@ -7877,7 +7877,7 @@ html[data-density='compact'] .status-badge {
   border: 1px solid var(--status-warning-border);
   border-radius: var(--radius-md);
   color: var(--status-warning-fg);
-  font-size: 12.5px;
+  font-size: 0.8125rem;
   line-height: 1.5;
 }
